Extract shared _CrossModalHFEmbedder base for SigLIP/SigLIP2/CLIP image embedders
The three HF cross-modal image embedders re-implemented the same
image/text embedding surface (~360 duplicated lines). The shared body
now lives in _cross_modal_shared.py; each subclass keeps only its
identity properties, its genuinely divergent _load_models_impl, a
_label for progress/log messages, and _text_processor_kwargs for the
per-backbone padding/truncation contract.

Refs #2651
