+++ vtscore/media/image/embedder_siglip2.py
@@ -7,17 +7,12 @@ the embedder loads on any ``transformers`` version that ships SigLIP 2 support
 
 from __future__ import annotations
 
-import logging
-from typing import TYPE_CHECKING, Any, Optional
-
-import numpy as np
+from typing import Any
 
 from vtscore.config import SIGLIP2_MODEL_ID
 from vtscore.media.embedder import (
     IMPORT_MODULE_ESTIMATES,
-    MediaEmbedder,
     embedder_load_setup,
-    extract_tensor as _extract_tensor,
     hf_token,
     intercept_tqdm_progress,
     intercept_weight_loading_progress,
@@ -25,13 +20,10 @@ from vtscore.media.embedder import (
     timed_progress,
     to_compute_device,
 )
-from vtscore.media.image._image_bulk import bulk_embed_image_files
-
-if TYPE_CHECKING:
-    from PIL import Image
+from vtscore.media.image._cross_modal_shared import _CrossModalHFEmbedder
 
 
-class ImageSiglip2Embedder(MediaEmbedder):
+class ImageSiglip2Embedder(_CrossModalHFEmbedder):
     """Embeds images using SigLIP 2 (google/siglip2-base-patch16-224).
 
     Successor to SigLIP with stronger transfer performance. Same 768-dim output
@@ -39,12 +31,8 @@ class ImageSiglip2Embedder(MediaEmbedder):
     queries, mirroring :class:`ImageSiglipEmbedder` so callers can swap.
     """
 
-    def __init__(self) -> None:
-        super().__init__()
-        # Typed ``Any``: transformers stubs miss several processor kwargs we
-        # pass at runtime; runtime ``None`` checks guard the calls.
-        self._model: Any = None
-        self._processor: Any = None
+    _label = "SigLIP 2"
+    _text_processor_kwargs: dict[str, Any] = {"padding": "max_length", "truncation": True}
 
     @property
     def name(self) -> str:
@@ -57,10 +45,6 @@ class ImageSiglip2Embedder(MediaEmbedder):
     @property
     def model_id(self) -> str:
         return SIGLIP2_MODEL_ID
-
-    @property
-    def media_type_id(self) -> str:
-        return "image"
 
     def _load_models_impl(self) -> None:
         if self._model is not None:
@@ -95,94 +79,6 @@ class ImageSiglip2Embedder(MediaEmbedder):
             self._processor = load_pretrained_local_first(
                 AutoProcessor.from_pretrained, SIGLIP2_MODEL_ID, cache_dir=cache_dir, token=hf_token()
             )
-
-    @property
-    def description_wrappers(self) -> list[str]:
-        return [
-            "a photo of {text}",
-            "a photograph of {text}",
-            "an image of {text}",
-            "{text}",
-            "a picture of {text}",
-        ]
-
-    def _embed_media_impl(self, media: dict) -> Optional[np.ndarray]:
-        if self._model is None:
-            self.load_models()
-        if self._model is None or self._processor is None:
-            return None
-        from vtscore.media.image._image_bulk import _load_pil, _pil_source_for  # noqa: PLC0415
-
-        source = _pil_source_for(media)
-        if source is None:
-            return None
-        image = _load_pil(source)
-        if image is None:
-            return None
-        return self.embed_pil_image(image)
-
-    def embed_pil_image(self, image: Image.Image) -> Optional[np.ndarray]:
-        if self._model is None:
-            self.load_models()
-        if self._model is None or self._processor is None:
-            return None
-        try:
-            import torch  # noqa: PLC0415
-
-            image = image.convert("RGB")
-            inputs = self._processor(images=image, return_tensors="pt")
-            device = next(self._model.parameters()).device
-            inputs = {k: v.to(device) for k, v in inputs.items()}
-            with torch.no_grad():
-                outputs = self._model.get_image_features(**inputs)
-                embedding = _extract_tensor(outputs).detach().cpu().numpy()
-            return embedding[0]
-        except Exception:
-            logging.getLogger(__name__).exception("Error embedding PIL image (SigLIP 2)")
-            return None
-
-    def _forward_pil_batch(self, images: list[Image.Image]) -> np.ndarray:
-        import torch  # noqa: PLC0415
-
-        rgb = [im.convert("RGB") for im in images]
-        inputs = self._processor(images=rgb, return_tensors="pt")
-        device = next(self._model.parameters()).device
-        inputs = {k: v.to(device) for k, v in inputs.items()}
-        with torch.no_grad():
-            outputs = self._model.get_image_features(**inputs)
-            return _extract_tensor(outputs).detach().cpu().numpy()
-
-    def _embed_media_bulk_impl(self, medias: list[dict]) -> list[Optional[np.ndarray]]:
-        if self._model is None:
-            self.load_models()
-        if self._model is None or self._processor is None:
-            return [None] * len(medias)
-        with self._embed_lock:
-            return bulk_embed_image_files(
-                medias,
-                forward_pil_batch=self._forward_pil_batch,
-                batch_size=self.embed_batch_size,
-                on_progress=self._on_progress,
-                label="SigLIP 2",
-            )
-
-    def _embed_text_impl(self, text: str) -> Optional[np.ndarray]:
-        if self._model is None:
-            self.load_models()
-        if self._model is None or self._processor is None:
-            return None
-        try:
-            import torch  # noqa: PLC0415
-
-            inputs = self._processor(text=[text], return_tensors="pt", padding="max_length", truncation=True)
-            device = next(self._model.parameters()).device
-            inputs = {k: v.to(device) for k, v in inputs.items()}
-            with torch.no_grad():
-                text_vec = _extract_tensor(self._model.get_text_features(**inputs)).detach().cpu().numpy()[0]
-            return text_vec
-        except Exception:
-            logging.getLogger(__name__).exception("Error embedding text query for image (SigLIP 2)")
-            return None
 
 
 EMBEDDER = ImageSiglip2Embedder()
